[#13] refactor: useChartDatas 구현
데이터 fetching과 가공, 로딩 상태와 에러 상태를 관리하는 훅 구현함
