+++ src/hooks/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { default as useChartDatas } from './useChartDatas';
